\name{surveyscores-package}


\alias{surveyscores-package}
\alias{surveyscores}

\docType{package}

\title{
  surveyscores
}

\description{
  Fonctions de calculs de scores de questionnaires depuis des echelles de reference.
}

\details{
  \tabular{ll}{
    Package: \tab  surveyscores \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/11/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Autres packages du meme auteur :

  \code{\link{lazyMe}}, \code{\link{dataDesc}}, \code{\link{manageR}},
  \code{\link{epiplus}}, \code{\link{sampleSize}}
}

\examples{

}
